Fix reading of nadir binary files
When reading more than one value into an array, we have to start reading
at the address of the first element of that array. Hence, the correct
call has to get array[0] as an argument which gets then converted to
&array[0] within the template binary_read(). This works the same way as
reading the radiance and error arrays a few lines below.

Changes the seconds in Nadir_Datensatz back to float. They are stored as
float in the binary files and we don't lose precision this way.

+++ LimbNadir_IO.h
@@ -42,7 +42,7 @@ public:
 	int m_Tag;
 	int m_Stunde;
 	int m_Minute;
-	int m_Sekunde;
+	float m_Sekunde;
 	float m_SZA_TOA[3];
 	float m_SAA_TOA[3];
 	float m_LOS_Zenit_Winkel[3];
